+++ CHANGELOG.md
@@ -9,8 +9,10 @@ The format is based on the [KeepAChangeLog] project.
 
 ### Fixed
 - [#711] Deal with no post_logout_redirect_uri
+- [#712] Set Content-Type on BackChannel logout POST.
 
 [#711]: https://github.com/OpenIDC/pyoidc/pull/711
+[#712]: https://github.com/OpenIDC/pyoidc/pull/712
 
 ## 1.1.1 [2019-11-04]